updating language to from ast grammar to grammar
