+++ src/grammar/mod.rs
@@ -1,3 +1,4 @@
+pub mod callable;
 pub mod expr;
 pub mod object;
 pub mod stmt;
