+++ src/main.rs
@@ -4,11 +4,11 @@ use getopts::Options;
 
 use crate::lox::Lox;
 
-mod ast_grammar;
 // mod ast_printer;
 // mod rpn_printer;
 mod environment;
 mod error;
+mod grammar;
 mod interpreter;
 mod lox;
 mod parser;
